feat : Add blueprint in app.py, api_fun.py, api.py.
